pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "any2json-py"
version = "0.1.0"
description = "Convert any file format into structured JSON using local parsers and LLMs"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "instructor>=1.0",
    "litellm>=1.40",
    "tiktoken>=0.7",
    "PyYAML>=6.0",
    "xmltodict>=0.13",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock>=3.14", "pytest-asyncio>=0.23"]

[project.scripts]
any2json-py = "any2json_py.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["any2json_py*"]
